InversionをShipStatusに移動
